fix prev merge error, implement gates

+++ app/Http/Controllers/api/v1/AddItemController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\api\v1;
 use App\Http\Controllers\Controller;
 use App\Models\AddItem;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AddItemController extends Controller  
 {
@@ -15,6 +16,10 @@ class AddItemController extends Controller
      */
     public function index()
     {
+        if(!Gate::allows(['admin', 'operator'])){
+            abort(403);
+        }
+
         $addItems = AddItem::latest()->get();
         return response([
             'success' => true,
@@ -31,6 +36,11 @@ class AddItemController extends Controller
     public function create()
     {
         //TODO: nunggu view create addItems
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
+        return "AddItem_create";
     }
 
     /**
@@ -70,6 +80,10 @@ class AddItemController extends Controller
      */
     public function show(AddItem $addItem)
     {
+        if(!Gate::allows(['admin', 'operator'])){
+            abort(403);
+        }
+
         // error_log(getrout);
         if ($addItem) {
             return response()->json([
@@ -95,6 +109,11 @@ class AddItemController extends Controller
     public function edit(AddItem $addItem)
     {
         //TODO: nunggu view edit addItem
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
+        return "AddItem_edit";
     }
 
     /**
@@ -138,6 +157,11 @@ class AddItemController extends Controller
      */
     public function destroy(AddItem $addItem)
     {
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
         $addItem->delete();
+        return "AddItem_destroy";
     }
 }
