+++ app/Http/Controllers/api/v1/SuperCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\api\v1;
 use App\Http\Controllers\Controller;
 use App\Models\SuperCategory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class SuperCategoryController extends Controller
 {
@@ -15,6 +16,10 @@ class SuperCategoryController extends Controller
      */
     public function index()
     {
+        if(!Gate::allows(['admin', 'operator'])){
+            abort(403);
+        }
+
         $superCategories = SuperCategory::latest()->get();
         return response([
             'success' => true,
@@ -31,6 +36,11 @@ class SuperCategoryController extends Controller
     public function create()
     {
         //
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
+        return "SuperCategory_create";
     }
 
     /**
@@ -62,6 +72,10 @@ class SuperCategoryController extends Controller
      */
     public function show(SuperCategory $superCategory)
     {
+        if(!Gate::allows(['admin', 'operator'])){
+            abort(403);
+        }
+
         if ($superCategory) {
             return response()->json([
                 'success' => true,
@@ -86,6 +100,11 @@ class SuperCategoryController extends Controller
     public function edit(SuperCategory $superCategory)
     {
         //
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
+        return "SuperCategory_edit";
     }
 
     /**
@@ -120,6 +139,11 @@ class SuperCategoryController extends Controller
      */
     public function destroy(SuperCategory $superCategory)
     {
+        if(!Gate::allows(['admin'])){
+            abort(403);
+        }
+
         $superCategory->delete();
+        return "SuperCategory_destroy";
     }
 }
